fix: add new years of catch if old doesn't exist increment year
Two problems in new catch, if the year didn't exist it wouldn't have
the same size matrix for the new additions, so, only negative years
are kept from the old file, if equilibrium catch was there, and all
values in the information matrix are rbinded to this equilibrium value
so it doesn't matter how many new years you are adding b/c it replaces
everything.

Also increments the last year of data to be the maximum year in the
catches.