feature: add tribal catches not in pacfin to pacfin data
Tribes are making an effort to use etickets, but until then they
cannot get their information uploaded in a timely manner for the
assessment. We told them previously that we were not going to include
this data if it wasn't in pacfin, but because they are making an effort
we decided in the 2020 assessment to include their numbers. The
catches are then included to weight the age composition information.