feature: write csv files of start date to disk
The stock assessment references the start date of the each sector and
the newly created files will allow for checking that the sectors did not
start fishing before the currently referenced date of May 15.
Files are written by sector, note that the additional sectors in the PacFIN
data correspond to non-hake midwater trawls and other non-hake sectors that
are not in the cooperatives. Dahl_sector 3 is the shoreside fishery.