feat(plot_boxplot): awesome figure function
Cleans up a lot of norpaccatches and plot_boxplot is now used instead
of box95. The major change is the use of ggplot here but the probabilities
are the same. The legacy function is kept inside of norpaccatches for reference
and because I think Chris might still be using it.

Much better determination of confidential or not confidential data.

Document the package via document(). This includes lots of old changes.