Simplify the number of functions and help fxns
Way too much code to review and things were getting lost.
Removed mydir that was internal in the functions for catches.
Now savedir is set outside as an argument call. Much better.
Decreased number of functions that need to be called by 1.
Fixes are through catch data only. Now working on compositions.