Adapt package to new structures 2024 assessment
This code now works again to pull, alter, and
save csv files in data-table. A major changes was the fetching of survey
files from the server rather than relying on the information in files we have
stored.
Emails were sent to Vanessa and Donna regarding found outliers in the weight-at-
age data.
An email will be sent to RT regarding 154 age-weight values that are in
out legacy files but not available on the network.